This should run SW and linear HW ramps
It has to be tested on the actual boards, but so far it compiles fine

+++ AD9954.h
@@ -96,7 +96,7 @@ class AD9954
 
     // places DDS in linear sweep mode
     //void linearSweep(unsigned long, unsigned long, unsigned long, byte, unsigned long, byte);
-    void linearSweep(unsigned long freq0, unsigned long freq1, float posTimeMicros, float negTimeMicros, float waitingTimeMicros, float powerPercentage, byte mode, bool noDwell = true);
+    void linearSweep(float freq0, float freq1, float posTimeMicros, float negTimeMicros, float waitingTimeMicros, float powerPercentage, byte mode, bool noDwell = true);
